Surviellence_camera/requirements.txt
flask
requests
opencv-python
numpy
pygame
playsound
paho-mqtt
opencv-contrib-python